Add re-usable implementation of DIA mock
