+++ pallets/oracle/src/oracle_mock.rs
@@ -86,9 +86,95 @@ impl Convert<u128, Option<FixedU128>> for MockConvertPrice {
 	}
 }
 
-pub struct MockConvertMoment;
-impl Convert<u64, Option<u64>> for MockConvertMoment {
-	fn convert(moment: u64) -> Option<u64> {
+pub struct MockConvertMoment<Moment>(sp_std::marker::PhantomData<Moment>);
+impl<Moment> Convert<Moment, Option<Moment>> for MockConvertMoment<Moment> {
+	fn convert(moment: Moment) -> Option<Moment> {
 		Some(moment)
+	}
+}
+
+// Implementation of re-usable mock DiaOracle
+
+thread_local! {
+	static COINS: RefCell<std::collections::HashMap<DataKey, Data>> = RefCell::new(std::collections::HashMap::<DataKey, Data>::new());
+}
+
+pub struct MockDiaOracle;
+impl dia_oracle::DiaOracle for MockDiaOracle {
+	fn get_coin_info(
+		blockchain: Vec<u8>,
+		symbol: Vec<u8>,
+	) -> Result<dia_oracle::CoinInfo, sp_runtime::DispatchError> {
+		let key = (blockchain, symbol);
+		let data_key = DataKey { blockchain: key.0.clone(), symbol: key.1.clone() };
+		let mut result: Option<Data> = None;
+		COINS.with(|c| {
+			let r = c.borrow();
+
+			let hash_set = &*r;
+			let o = hash_set.get(&data_key);
+			match o {
+				Some(i) => result = Some(i.clone()),
+				None => {},
+			};
+		});
+		let Some(result) = result else {
+			return Err(sp_runtime::DispatchError::Other(""));
+		};
+		let mut coin_info = dia_oracle::CoinInfo::default();
+		coin_info.price = result.price;
+		coin_info.last_update_timestamp = result.timestamp;
+
+		Ok(coin_info)
+	}
+
+	//Spacewalk DiaOracleAdapter does not use get_value function. There is no need to implement
+	// this function.
+	fn get_value(
+		_blockchain: Vec<u8>,
+		_symbol: Vec<u8>,
+	) -> Result<dia_oracle::PriceInfo, sp_runtime::DispatchError> {
+		unimplemented!(
+			"DiaOracleAdapter implementation of DataProviderExtended does not use this function."
+		)
+	}
+}
+
+use orml_oracle::{DataFeeder, DataProvider, TimestampedValue};
+use std::cell::RefCell;
+
+pub struct MockDataCollector<AccountId, Moment>(
+	sp_std::marker::PhantomData<AccountId>,
+	sp_std::marker::PhantomData<Moment>,
+);
+
+//DataFeeder required to implement DataProvider trait but there no need to implement get function
+impl<AccountId, Moment> DataProvider<Key, TimestampedValue<UnsignedFixedPoint, Moment>>
+	for MockDataCollector<AccountId, Moment>
+{
+	fn get(_key: &Key) -> Option<TimestampedValue<UnsignedFixedPoint, Moment>> {
+		unimplemented!("Not required to implement DataProvider get function")
+	}
+}
+impl<AccountId, Moment: Into<u64>>
+	DataFeeder<Key, TimestampedValue<UnsignedFixedPoint, Moment>, AccountId>
+	for MockDataCollector<AccountId, Moment>
+{
+	fn feed_value(
+		_who: AccountId,
+		key: Key,
+		value: TimestampedValue<UnsignedFixedPoint, Moment>,
+	) -> sp_runtime::DispatchResult {
+		let key = MockOracleKeyConvertor::convert(key).unwrap();
+		let r = value.value.into_inner();
+
+		let data_key = DataKey { blockchain: key.0.clone(), symbol: key.1.clone() };
+		let data = Data { key: data_key.clone(), price: r, timestamp: value.timestamp.into() };
+
+		COINS.with(|coins| {
+			let mut r = coins.borrow_mut();
+			r.insert(data_key, data);
+		});
+		Ok(())
 	}
 }
